+++ src/utility_funcs.rs
@@ -1,6 +1,6 @@
-const PI: f64 = 3.1415926535897932385;
+const PI: f32 = 3.1415926535897932385;
 
-pub fn clamp(x: f64, min: f64, max:f64) -> f64 {
+pub fn clamp(x: f32, min: f32, max:f32) -> f32 {
     match x {
         x if x < min => min,
         x if x > max => max,
@@ -8,6 +8,6 @@ pub fn clamp(x: f64, min: f64, max:f64) -> f64 {
     }
 }
 
-pub fn degrees_to_radians(degrees: f64) -> f64 {
+pub fn degrees_to_radians(degrees: f32) -> f32 {
     degrees * PI/180.0
 }
